requirements.txt
aiogram==3.4.1
groq==0.4.2
sqlalchemy==2.0.25
python-dotenv==1.0.1
gspread==6.0.0
google-auth==2.27.0
apscheduler==3.10.4